requirements.txt
diffusers==0.33.1
transformers==4.53.0
accelerate==1.8.1
matplotlib
torchmetrics==1.7.3
protobuf==6.31.1
sentencepiece==0.2.0